Make uninstall work when not installed

+++ src/DoctrineTermStore.php
@@ -40,16 +40,22 @@ class DoctrineTermStore implements TermStore {
 			$reporter = new NullMessageReporter();
 		}
 
-		$reporter->reportMessage( 'Uninstalling Wikibase Term Store: removing tables... ' );
-
 		$schema = $this->connection->getSchemaManager();
-		$schema->dropTable( $this->tableNames->itemTerms() );
-		$schema->dropTable( $this->tableNames->propertyTerms() );
-		$schema->dropTable( $this->tableNames->termInLanguage() );
-		$schema->dropTable( $this->tableNames->textInLanguage() );
-		$schema->dropTable( $this->tableNames->text() );
 
-		$reporter->reportMessage( "done\n" );
+		if ( $schema->tablesExist( $this->tableNames->itemTerms() ) ) {
+			$reporter->reportMessage( 'Uninstalling Wikibase Term Store: removing tables... ' );
+
+			$schema->dropTable( $this->tableNames->itemTerms() );
+			$schema->dropTable( $this->tableNames->propertyTerms() );
+			$schema->dropTable( $this->tableNames->termInLanguage() );
+			$schema->dropTable( $this->tableNames->textInLanguage() );
+			$schema->dropTable( $this->tableNames->text() );
+
+			$reporter->reportMessage( "done\n" );
+		}
+		else {
+			$reporter->reportMessage( 'Wikibase Term Store is not installed so will not be removed' );
+		}
 	}
 
 	public function newPropertyTermStore(): PropertyTermStore {
